refacto/global: review all app to refacto
