+++ src/components/ModalCreationChambre.tsx
@@ -150,7 +150,7 @@ export default function ModalCreationChambre({
           </form>
         </div>
 
-        {/* Footer — boutons */}
+        {/* Footer boutons */}
         <div className="flex gap-3 px-6 py-4 border-t border-border shrink-0">
           <button
             type="button"
